FidelityGenerativeAINanoDegree/project02_OpenAI_Chatbot_with_RAG/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
tiktoken>=0.5.1
wikipedia>=1.4.0
beautifulsoup4>=4.12.2
numpy>=1.24.1